+++ tests/test_distros.py
@@ -99,7 +99,7 @@ class DistroTestsBase(MoncicTestCase, abc.ABC):
             with self.subTest(parsed=parsed):
                 self.assertIs(DistroFamily.from_osrelease(parsed, "invalid"), self.distro)
 
-                with tempfile.TemporaryDirectory() as root_str:
+                with tempfile.TemporaryDirectory(suffix="test_from_osrelease") as root_str:
                     root = Path(root_str)
                     path = root / "etc" / "os-release"
                     path.parent.mkdir(parents=True)
@@ -121,7 +121,6 @@ class DistroTestsBase(MoncicTestCase, abc.ABC):
         path = self.workdir()
         self.distro.bootstrap(session.bootstrapper, path)
         self.assertBootstrapCommands(session.run_log, path)
-        self.assertRunLogEmpty(session.run_log)
 
     def test_update(self) -> None:
         session = self.session()
@@ -134,10 +133,8 @@ class DistroTestsBase(MoncicTestCase, abc.ABC):
         bootstrapped.update()
 
         with self.match_run_log(session.run_log) as m:
-            container_run_log = m.assertPopFirst(f"{self.distro.full_name}: container start")
+            container_run_log = m.assertPopFirst(f"{self.distro.full_name}: run container")
             self.assertUpdateCommands(container_run_log, Path("/test"))
-            m.assertPopFirst(f"{self.distro.full_name}: container stop")
-            m.assertEmpty()
 
 
 class TestCentos7(DistroTestsBase):
@@ -200,7 +197,6 @@ class DebDistroTestsBase(DistroTestsBase):
         ]
         with self.match_run_log(run_log) as m:
             m.assertPopFirst(re.compile(" ".join(match)))
-            m.assertEmpty()
 
     @override
     def assertUpdateCommands(self, run_log: RunLogEntry, path: Path) -> None: